% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_features.R
\name{add_features}
\alias{add_features}
\title{Add features to ESM data}
\usage{
add_features(esm, specifications, db)
}
\arguments{
\item{db}{}
}
\description{
Add features to ESM data
}
\details{
\enumerate{
\item \code{\link[=identify_gaps]{identify_gaps()}}: Identifying gaps in the data (if needed).
\item \code{\link[=link_gaps]{link_gaps()}}: Linking gaps to the ESM data (if needed).
\item \code{\link[=feature_data]{feature_data()}}: Extracting feature-specific data for each sensor.
\item \code{\link[=add_gaps]{add_gaps()}}: Adding gaps to the feature data (if needed).
\item \code{\link[=link]{link()}}: Linking the feature data to the ESM data within a time window.
\item Various calls to sensor-specific functions, such as \code{\link[=activity_duration]{activity_duration()}},
\code{\link[=noise_intensity]{noise_intensity()}}, or \code{\link[=step_count]{step_count()}}.
}
}
