% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\name{add_sensor_spec}
\alias{add_sensor_spec}
\title{Add sensor-specific arguments to a specification list}
\usage{
add_sensor_spec(
  specifications,
  sensor,
  preprocess_bin = FALSE,
  preprocess_moving_average = FALSE,
  preprocess_tz = Sys.timezone(),
  gaps = TRUE,
  gaps_continue = FALSE,
  gaps_fill = NULL,
  link_add_before = TRUE,
  link_add_after = TRUE,
  ...
)
}
\arguments{
\item{specifications}{A specific list as created by \code{\link[=create_specs]{create_specs()}}.}

\item{sensor}{The name of a sensor. See \link[mpathsenser]{sensors} for a list of available sensors.}

\item{preprocess_bin}{Whether raw sensor data should be binned to 1-minute intervals before
processing. Corresponds to the \code{bin} argument in \code{\link[=feature_data]{feature_data()}}.}

\item{preprocess_moving_average}{Whether a moving average should be calculated for the raw sensor
data before processing. Corresponds to the \code{moving_average} argument in \code{\link[=feature_data]{feature_data()}}.}

\item{preprocess_tz}{The time zone to be used in processing the data. See \code{\link[=OlsonNames]{OlsonNames()}} for a
full list of the available time zones.}

\item{gaps}{Should gaps be included?}

\item{gaps_continue}{Whether to continue the measurement(s) prior to the gap once the gap ends.
Corresponds to the \code{continue} argument in \code{\link[=add_gaps]{add_gaps()}}.}

\item{gaps_fill}{A named list of the columns to fill with default values for the extra
measurements that are added because of the gaps. Corresponds to the \code{fill} argument in
\code{\link[=add_gaps]{add_gaps()}}.`}

\item{link_add_before}{Whether to add the last measurement before the start of each interval.
Corresponds to the \code{add_before} argument in \code{\link[=link]{link()}}.}

\item{link_add_after}{Whether to add the first measurement after the end of each interval.
Corresponds to the \code{add_after} argument in \code{\link[=link]{link()}}.}

\item{...}{}
}
\value{
A list of specifications
}
\description{
Add sensor-specific arguments to a specification list
}
