% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CARP-package.R
\docType{package}
\name{CARP-package}
\alias{CARP}
\alias{CARP-package}
\title{CARP: Process and Analyse Data from CARP Mobile Sensing}
\description{
Process and Analyse Data from CARP Mobile Sensing.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://gitlab.kuleuven.be/ppw-okpiv/researchers/u0134047/CARP/issues}
}

}
\author{
\strong{Maintainer}: Koen Niemeijer \email{koen.niemeijer@kuleuven.be} (\href{https://orcid.org/0000-0002-0816-534X}{ORCID})

}
\keyword{internal}
