% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean}
\alias{clean}
\title{Clean up direct CARP output}
\usage{
clean(path = getwd(), zipped = TRUE, fix = FALSE, overwrite = FALSE)
}
\arguments{
\item{path}{The path to the directory containing the JSON and/or zip files}

\item{zipped}{Logical value indicating whether you want to unzip the files as well.}

\item{fix}{Logical value indicating whether to fix incorrectly formatted JSON files in path.}

\item{overwrite}{Defaults to true.}
}
\value{
Invisible result. Side effect is corrupted JSON files fixed and/or data unzipped.
}
\description{
When copying data directly coming from CARP, JSON files are often corrupt due to the
app not properly closing them. This function (hopefully) fixes that. Additionally,
it can also unzip all the output in the directory.
}
\details{
Note: Be careful when running this function multiple times. In the case there are
many zips and no JSON files, it will be fast. However, when re-running the function,
every previously unzipped JSON file will be checked for errors.
}
