% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_carp}
\alias{clean_carp}
\title{Clean up direct CARP output}
\usage{
clean_carp(path = getwd(), zipped = TRUE)
}
\arguments{
\item{path}{The path to the directory containing the JSON and/or zip files}

\item{zipped}{Logical value indiciating whether you want to unzip the files as well.}
}
\value{
Invisible result. Side effect is corrupted JSON files fixed and/or data unzipped.
}
\description{
When copying data directly coming from CARP, JSON files are often corrupt due to the
app not properly closing them. This function (hopefully) fixes that. Additionally,
it can also unzip all the output in the directory.
}
