% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{close_db}
\alias{close_db}
\title{Close a database connection}
\usage{
close_db(db)
}
\arguments{
\item{db}{A database connection to an m-Path Sense database.}
}
\value{
\code{close_db} returns invisibly regardless of whether the database is active, valid,
or even exists.
}
\description{
This is a convenience function that is simply a wrapper around \link[DBI]{dbDisconnect}.
}
\seealso{
\code{\link[mpathsenser]{open_db}} for opening an mpathsenser database.
}
