% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{collect.coverage}
\alias{collect.coverage}
\title{Materialise a coverage result}
\usage{
\method{collect}{coverage}(x, ...)
}
\arguments{
\item{x}{A lazy coverage table created by \code{\link[=coverage_frequency]{coverage_frequency()}} or
\code{\link[=coverage_proportional]{coverage_proportional()}}.}

\item{...}{Passed on to the underlying \code{collect()} method.}
}
\value{
A tibble with rounded coverage values and factorised \code{measure}.
}
\description{
Materialise a coverage result
}
\keyword{internal}
