% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\name{create_specs}
\alias{create_specs}
\title{Create a default specification list}
\usage{
create_specs(
  time_window_before,
  time_window_after,
  min_gap = 60L,
  gap_sensors = sensors,
  tz = Sys.timezone(),
  defaults = NULL,
  sensors = NULL,
  ...
)
}
\arguments{
\item{time_window_before}{The time period before each beep, as in \code{\link[=link]{link()}}.}

\item{time_window_after}{The time period after each beep, as in \code{\link[=link]{link()}}.}

\item{min_gap}{he minimum time (in seconds) passed between two subsequent measurements for it to
be considered a gap, as in \code{\link[=identify_gaps]{identify_gaps()}}.}

\item{gap_sensors}{One or multiple \code{\link[mpathsenser]{sensors}}, as in \code{\link[=identify_gaps]{identify_gaps()}}.
See sensors for a list of available sensors.}

\item{tz}{The time zone to be used in processing the data. See \code{\link[=OlsonNames]{OlsonNames()}} for a full list of
the available time zones.}

\item{...}{Arguments passed on to \code{\link[=add_sensor_spec]{add_sensor_spec()}}.}
}
\value{
A list containing sensor-specific arguments and time_window arguments. See the details
section for a more precise description.
}
\description{
Create a default specification list
}
\examples{
create_specs(1800, 0)
#'
# To modify a single parameter in a list of specifications,
# use purrr's assign_in:
specs <- list(create_specs(1800L, 0), create_specs(0, 1800))
purrr::map(.x = specs,
           .f = purrr::assign_in,
           where = c("gyroscope", "gaps", "gaps"),
           value = TRUE)

# Alternatively, it may be added to replace the sensor entirely
# if many values need to be changed. This is possible using
# add_sensor_spec() (note it replaces all values for that sensor):
purrr::map(.x = specs,
           .f = add_sensor_spec,
           sensor = "gyroscope",
           preprocess_bin = FALSE,
           preprocess_moving_average = FALSE,
           gaps = TRUE,
           gaps_continue = TRUE,
           link_add_after = TRUE,
           link_add_before = TRUE)
}
