% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_data.R
\name{feature_data.default}
\alias{feature_data.default}
\alias{feature_data.accelerometer}
\alias{feature_data.activity}
\alias{feature_data.battery}
\alias{feature_data.bluetooth}
\alias{feature_data.gyroscope}
\alias{feature_data.location}
\alias{feature_data.noise}
\title{Extract cleaned data from an mpathsenser database}
\usage{
\method{feature_data}{default}(
  db,
  sensor,
  cols = NULL,
  ...,
  bin = FALSE,
  moving_average = NULL,
  tz = Sys.timezone(),
  call = NULL
)

\method{feature_data}{accelerometer}(
  db,
  sensor = "Accelerometer",
  cols = NULL,
  ...,
  bin = TRUE,
  moving_average = NULL,
  tz = Sys.timezone()
)

\method{feature_data}{activity}(db, sensor = "Activity", cols = NULL, ..., tz = Sys.timezone())

\method{feature_data}{battery}(db, sensor = "Battery", cols = NULL, ..., tz = Sys.timezone())

\method{feature_data}{bluetooth}(db, sensor = "Bluetooth", cols = NULL, ..., tz = Sys.timezone())

\method{feature_data}{gyroscope}(
  db,
  sensor = "Gyroscope",
  cols = NULL,
  ...,
  bin = FALSE,
  moving_average = NULL,
  tz = Sys.timezone()
)

\method{feature_data}{location}(
  db,
  sensor = "Location",
  cols = NULL,
  ...,
  key = NULL,
  cache = NULL,
  tz = Sys.timezone()
)

\method{feature_data}{noise}(
  db,
  sensor = "Noise",
  cols = NULL,
  ...,
  bin = FALSE,
  moving_average = NULL,
  tz = Sys.timezone()
)
}
\arguments{
\item{db}{A database connection to an m-Path Sense database.}

\item{sensor}{The name of a sensor. See \link[mpathsenser]{sensors} for a list of available
sensors.}

\item{cols}{One or multiple column names to extract for this sensor. These names must correspond
to the ones in the database. If left empty, all variables will be extracted.}

\item{bin}{Whether to bin the data. Only 1-minute bins are possible.}

\item{moving_average}{The number of seconds to calculate a moving average over. Defaults to no
moving average.}

\item{tz}{The time zone used to convert the time column to. In the database this defaults to UTC
but this function uses the user's current time zone as this is likely the same time zone as the
data was collected in.}

\item{call}{Internal use only. A call to execute before \link[dplyr:compute]{dplyr::collect} is called. Note that
these columns must carry the symbol \code{data} as their target data frame.}

\item{key}{A decryption key string. Only applicable when extraction location data.}
}
\value{
A \link[tibble:tibble]{tibble::tibble} containing the columns \code{participant_id}, \code{time}, and other specified
columns.
}
\description{
Extract cleaned data from an mpathsenser database
}
\keyword{internal}
