% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_data.R
\name{feature_data}
\alias{feature_data}
\title{Extract cleaned data from an mpathsenser database}
\usage{
feature_data(db, sensor, cols = NULL, ..., tz = Sys.timezone())
}
\arguments{
\item{db}{A database connection to an m-Path Sense database.}

\item{sensor}{The name of a sensor. See \link[mpathsenser]{sensors} for a list of available
sensors.}

\item{cols}{One or multiple column names to extract for this sensor. These names must correspond
to the ones in the database. If left empty, all variables will be extracted.}

\item{tz}{The time zone used to convert the time column to. In the database this defaults to UTC
but this function uses the user's current time zone as this is likely the same time zone as the
data was collected in.}

\item{bin}{Whether to bin the data. Only 1-minute bins are possible.}

\item{moving_average}{The number of seconds to calculate a moving average over. Defaults to no
moving average.}

\item{min_confidence}{Activity data only. The minimum confidence value needed to keep the measurement. See \href{https://developers.google.com/android/reference/com/google/android/gms/location/DetectedActivity#getConfidence()}{Android's documentation} for a more precise description of what the confidence value entails.}

\item{key}{A decryption key string. Only applicable when extraction location data.}
}
\value{
A \link{tibble} containing the columns \code{participant_id}, \code{time}, and other specified
columns.
}
\description{
Extract cleaned data from an mpathsenser database
}
