% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fix_json}
\alias{fix_json}
\title{Fix JSON files}
\usage{
fix_json(path = getwd())
}
\arguments{
\item{path}{The pathname of the JSON files.}
}
\description{
When copying data directly coming from CARP, JSON files are often corrupt due to the
app not properly closing them. This function attempts to fix the most common
problems associated with improper file closure by CARP.
}
\details{
Note: Be careful when running this function multiple times. In the case there are
many zips and no JSON files, it will be fast. However, when re-running the function,
every previously unzipped JSON file will be checked for errors.
}
