% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{freq}
\alias{freq}
\title{Measurement frequencies per sensor}
\usage{
freq
}
\value{
This vector contains the following
information:\tabular{lll}{
   Sensor \tab Frequency (per hour) \tab Full text \cr
   Accelerometer \tab 720 \tab Once per 5 seconds. Can have multiple instances. \cr
   Activity \tab 120 \tab Once per 30 seconds. \cr
   AppUsage \tab 2 \tab Once every 30 minutes. Can have multiple instances. \cr
   Battery \tab 60 \tab Once per minute. \cr
   Bluetooth \tab 12 \tab Once every 5 minutes. Can have multiple instances. \cr
   BluetoothBeacon \tab 12 \tab Once every 5 minutes. Can have multiple instances. \cr
   Connectivity \tab 12 \tab Once every 5 minutes. \cr
   Device \tab 60 \tab Once per minute. \cr
   Error \tab 12 \tab Once every 5 minutes. \cr
   Heartbeat \tab 60 \tab Once per minute. \cr
   Light \tab 360 \tab Once per 10 seconds. \cr
   Location \tab 60 \tab Once every 60 seconds. \cr
   Memory \tab 60 \tab Once per minute \cr
   Pedometer \tab 60 \tab Once per minute. \cr
   Screen \tab 12 \tab Once every 5 minutes. \cr
   Timezone \tab 12 \tab Once every 5 minutes. \cr
   Weather \tab 1 \tab Once per hour. \cr
   Wifi \tab 60 \tab Once per minute. \cr
}
}
\description{
A numeric vector containing (an example) of example measurement frequencies per sensor.
Such input is needed for \code{\link[=coverage]{coverage()}}.
}
\examples{
freq
}
