% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_functions.R
\name{get_app_usage}
\alias{get_app_usage}
\title{Get app usage per hour}
\usage{
get_app_usage(db, participant_id = NULL, startDate = NULL, endDate = NULL)
}
\arguments{
\item{db}{A database connection to a CARP database.}

\item{participant_id}{A character string identifying a single participant. Use
\code{\link[CARP]{get_participants}} to retrieve all participants from the database.
Leave empty to get data for all participants.}

\item{startDate}{Optional search window specifying date where to begin search. Must be convertible to date using \link[base]{as.Date}. Use \link[CARP]{first_date} to find the date of the first entry for a participant.}

\item{endDate}{Optional search window specifying date where to end search. Must be convertible to date using \link[base]{as.Date}. Use \link[CARP]{last_date} to find the date of the last entry for a participant.}
}
\value{
A data frame containing a column "App" and a column "Usage" for the hourly app usage.
}
\description{
This function extracts app usage per hour for either one or multiple participants. If multiple
days are selected, the app usage time is averaged.
}
