% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{index_db}
\alias{index_db}
\title{Create indexes for a mpathsenser database}
\usage{
index_db(db)
}
\arguments{
\item{db}{A database connection to an m-Path Sense database.}
}
\value{
No return value, called for side effects.
}
\description{
Create indexes for a mpathsenser database
}
