% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_functions.R
\name{installed_apps}
\alias{installed_apps}
\title{Get installed apps}
\usage{
installed_apps(db, participant_id = NULL)
}
\arguments{
\item{db}{A database connection to an mpathsenser database.}

\item{participant_id}{A single participant identifier (stored as an unsigned integer; an
integer, numeric, or character value is accepted). Use
\code{\link[mpathsenser]{get_participants}} to retrieve all participants from the database.
Leave empty to get data for all participants.}
}
\value{
A tibble containing app names.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Extract installed apps for one or all participants. Contrarily to other get_* functions in
this package, start and end dates are not used since installed apps are assumed to be fixed
throughout the study.
}
\examples{
\dontrun{
db <- example_db()

# Get installed apps for all participants
installed_apps(db)

# Get installed apps for a single participant
installed_apps(db, 372780)
}
}
