% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purge_redundant.R
\name{purge_redundant}
\alias{purge_redundant}
\title{Purge redundant rows that are preceded and/or succeeded by rows with the same values in selected
columns}
\usage{
purge_redundant(.data, .cols = NULL, .before = TRUE, .after = TRUE)
}
\arguments{
\item{.data}{A data frame containing the data to be compressed}

\item{.cols}{A selection of columns in which to check for consecutive identical values. If none
are specified, all columns will be used.}

\item{.before}{A logical value indicating whether to check for identical values preceding each
row. Defaults to TRUE.}

\item{.after}{A logical value indicating whether to check for identical values succeeding each
row. Defaults to TRUE.}
}
\value{
A data frame with rows containing consecutive identical values in the specified columns
removed.
}
\description{
Purge redundant rows that are preceded and/or succeeded by rows with the same values in selected
columns
}
\examples{
data <- data.frame(time = c("2022-01-01", "2022-01-02", "2022-01-03",
                           "2022-01-03", "2022-01-03", "2022-01-05"),
                  x = c(5, 4, 3, 3, 3, 2))
purge_redundant(data, c(x, y))
}
