% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{split_db}
\alias{split_db}
\title{Split a subset of a database into another database}
\usage{
split_db(old_db, new_db = NULL, sensor, path = getwd(), db_name = "carp.db")
}
\arguments{
\item{old_db}{A CARP database connection from where the data will be transferred.}

\item{new_db}{A CARP database connection where the data will be transferred to. If no new_db is
specified, a path (and possibly a db_name) must be specified for \link[CARP]{create_db} to
create a new database.}

\item{sensor}{A character vector containing one or multiple sensors. See
\code{\link[CARP]{sensors}} for a list of available sensors. Use "All" for all available sensors.}

\item{path}{The path to the database. Use NULL to use the full path name in db_name.}

\item{db_name}{The name of the database.}
}
\description{
Split a subset of a database into another database
}
