% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{total_acceleration}
\alias{total_acceleration}
\title{Calculate total acceleration from x, y, and z coordinates}
\usage{
total_acceleration(data, colname, x = x, y = y, z = z, gravity = 9.810467)
}
\arguments{
\item{data}{A data frame or a remote data frame connection through \link[dplyr]{tbl}.}

\item{colname}{The name of the newly added column containing the total acceleration.}

\item{x}{Acceleration along the x-axis.}

\item{y}{Acceleration along the y-axis.}

\item{z}{Acceleration along the z-axis.}

\item{gravity}{Gravity in meters per second. Defaults to 9.810467. Set to 0 to ignore.}
}
\value{
The input data with a column \code{colname} attached.
}
\description{
A convenience function to calculate the total acceleration in either or a local or remote
tibble.
}
\examples{
\dontrun{
db <- open_db()
tbl(db, "Accelerometer") \%>\%
  total_acceleration("total_acc")
}
}
