% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{unzip_carp}
\alias{unzip_carp}
\title{Unzip CARP output}
\usage{
unzip_carp(path = getwd(), overwrite = FALSE)
}
\arguments{
\item{path}{The path to the directory containing the zip files.}

\item{overwrite}{Logical value wheter you want to overwrite already existing zip files.}
}
\description{
Similar to \link[utils]{unzip}, but makes it easier to unzip all files in a given path
with one function call.
}
