% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_data}
\alias{unzip_data}
\title{Unzip m-Path Sense output}
\usage{
unzip_data(
  path,
  to = NULL,
  overwrite = FALSE,
  recursive = TRUE,
  .progress = TRUE
)
}
\arguments{
\item{path}{The path to the directory containing the zip files.}

\item{to}{The output path. Defaults to \code{path}.}

\item{overwrite}{Logical value whether you want to overwrite already existing zip files.}

\item{recursive}{Logical value indicating whether to search subdirectories recursively.
Extracted files are placed in the corresponding subdirectory of \code{to}.}

\item{.progress}{Whether to display a progress bar.}
}
\value{
Invisibly returns the number of ZIP files successfully unzipped.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Similar to \link[utils]{unzip}, but makes it easier to unzip all files in a given path with one
function call.
}
\examples{
\dontrun{
# Unzip all files in a directory
unzip_data(path = "path/to/zipfiles", to = "path/to/unzipped", recursive = FALSE)

# Unzip files recursively
unzip_data(path = "path/to/zipfiles", to = "path/to/unzipped", recursive = TRUE)

# Skip files that are already unzipped
unzip_data(path = "path/to/zipfiles", to = "path/to/unzipped", overwrite = FALSE)
}
}
