% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifications.R
\name{validate_specifications}
\alias{validate_specifications}
\title{Check that object is a valid specifications list}
\usage{
validate_specifications(
  specifications,
  arg = rlang::caller_arg(specifications),
  call = rlang::caller_env()
)
}
\arguments{
\item{specifications}{A specifications object, as created by \code{\link[=create_specs]{create_specs()}}.}

\item{arg}{An argument name as a string. This argument will be mentioned in error messages as the
input that is at the origin of a problem.}

\item{call}{The execution environment of a currently running function.}
}
\value{
Returns \code{TRUE} invisibly.
}
\description{
Check that object is a valid specifications list
}
