% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_functions.R
\name{wifi_home_ssid}
\alias{wifi_home_ssid}
\title{Find the home network SSID}
\usage{
wifi_home_ssid(data, exclude = "GAP")
}
\arguments{
\item{data}{A data frame containing the Wi-Fi data with at least the columns \code{time} and \code{ssid}.}

\item{exclude}{Optionally, a character vector of values to exclude from the SSID list. This is
particularly useful if gaps have already been added to the data using \code{\link[=add_gaps]{add_gaps()}}.}
}
\value{
A \code{tibble} of home network SSIDs, grouped according to \code{data}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

One way of measuring time spent at home is to measure the time connected to the home Wi-Fi
network (see \code{\link[=wifi_time_home]{wifi_time_home()}}). To do this, you first need to which network the home Wi-Fi
network is. \code{wifi_home_ssid()} attempts to find the most likely candidate to be the home Wi-Fi
network.
}
\details{
It is not an easy task to determine which SSID (service set ID) from a set of (hashed)
SSIDs is most likely someone's home network. One way to do this is by combining it with GPS
data, but there we face the same problem in that we do not know where someone's "home" is
(unless it is asked somehow). Thus, this function solves the problem in two steps:
\enumerate{
\item Find the most often occurring (connected) network SSID between 12AM and 6AM, when people are
generally home the most.
\item If no result was found (e.g. because of airplane mode, phone turned
off, etcetera), take the most often occurring SSID throughout the entire day.
}
}
\section{SSID vs BSSID vs IP}{
 You may have noticed that in the mpathsenser database there are
entries in the Wi-Fi table for SSIDs, BSSIDs, and IP addresses. However, this function focusses
solely on SSIDs.

The IP addresses refer to the interal IP address of the network the person is currently
connected to. Consequently, most values are actually empty as only Wi-Fi networks that the
person is currently connected to have an IP address. Moreover, these are internal IP addresses,
so they may be replicated throughout disparate networks and within the same network, the
internal IP address may change over time. Consequently, IP addresses are not used in this
function.

As for BSSIDs basic service set identifier (BSSID), the relationship between SSIDs and BSSIDs
is that one SSIDs may be associated with multiple BSSIDs, as there are unique to an access
point (e.g. a router).Thus, if a home contains multiple access points, it will have multiple
BSSIDs.
}

\seealso{
Other Wi-Fi functions: 
\code{\link{wifi_recurring_ssids}()}
}
\concept{Wi-Fi functions}
