% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_functions.R
\name{wifi_recurring_ssids}
\alias{wifi_recurring_ssids}
\title{Find recurring network SSIDs}
\usage{
wifi_recurring_ssids(data, n = 10, exclude = "GAP")
}
\arguments{
\item{data}{A data frame containing the Wi-Fi data with at least the columns \code{time} and \code{ssid}.}

\item{n}{The number of recurring SSIDs to extract.}

\item{exclude}{Optionally, a character vector of values to exclude from the SSID list. This is
particularly useful if gaps have already been added to the data using \code{\link[=add_gaps]{add_gaps()}}.}
}
\value{
A \code{tibble} of recurring network SSIDs, grouped according to \code{data}.
}
\description{
To provide estimate of how often someone revisits places, you must first find out which
"significant" places exist. In a similar vein as \code{\link[=location_unique_places]{location_unique_places()}}, this function
extract the most often recurring "places" through the associated network SSIDs (assuming most
places have Wi-Fi networks in the vicinity).
}
\seealso{
Other Wi-Fi functions: 
\code{\link{wifi_home_ssid}()}
}
\concept{Wi-Fi functions}
