% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature-implementations.R
\name{activity_duration}
\alias{activity_duration}
\title{Activity duration feature}
\usage{
activity_duration(
  data,
  weight_by = NULL,
  gap_value = "GAP",
  categories = c("ACTIVE", "STILL", "UNKNOWN")
)
}
\arguments{
\item{data}{The activity data}

\item{weight_by}{A single column name in \code{data} to weigh the observations by.}
}
\description{
Activity duration feature
}
\concept{activity}
