% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_functions.R
\name{add_gaps}
\alias{add_gaps}
\title{Add gap periods to sensor data}
\usage{
add_gaps(data, gaps, by = NULL, fill = NULL)
}
\arguments{
\item{data}{A data frame containing the activity data. See \link[mpathsenser]{get_data} for
retrieving activity data from an mpathsenser database.}

\item{gaps}{A data frame (extension) containing the gap data. See
\link[mpathsenser]{identify_gaps} for retrieving gap data from an mpathsenser database. It
should at least contain the columns \code{from} and \code{to} (both in a date-time format), as
well as any specified columns in \code{by}.}

\item{by}{A character vector indicating the variable(s) to match by, typically the participant
IDs. If NULL, the default, \code{*_join()} will perform a natural join, using all variables in
common across \code{x} and \code{y}.}

\item{fill}{A named list of the columns to fill with default values for the extra measurements
that are added because of the gaps.}
}
\value{
A tibble containing the data and the added gaps.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
