% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alias_column_names.R
\name{alias_column_names}
\alias{alias_column_names}
\title{Alias column names of a data frame containing sensor data}
\usage{
alias_column_names(data, ...)
}
\arguments{
\item{data}{A data frame for which the column names are remapped according to the specific method
for this data frame. As such, \code{data} should have a class attribute that corresponds to one of
the methods defined in this file.}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A data frame with the column names remapped according to the specific method for this
data frame.
}
\description{
This function is used to remap the column names of a data frame containing sensor data to the
column names used in the database schema of mpathsenser. As the column names have already
changed several times over the year, we use these functions so we can easily add new aliases of
the existing columns. The function is generic and dispatches to the specific method for the
class of \code{data}.
}
\details{
Aliasing of the column names depends on the specific method, but column names are only changed
if they have an alias available. Otherwise, they are kept unchanged.
}
\keyword{internal}
