% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_functions.R
\name{app_category}
\alias{app_category}
\title{Find the category of an app on the Google Play Store}
\usage{
app_category(name, num = 1)
}
\arguments{
\item{name}{The name of the app to search for.}

\item{num}{Which result should be selected in the list of search results. Defaults to one.}
}
\value{
A list containing the following fields:\tabular{ll}{
    \tab  \cr
   package \tab the package name that was selected from the Google Play search \cr
   genre \tab the corresponding genre of this package \cr
}
}
\description{
This function scrapes the Google Play Store by using \code{name} as the search term. From there
it selects the first result in the list and its corresponding category and package name.
}
\examples{
app_category("whatsapp")

# Example of a generic app name where we can't find a specific app
app_category("weather") # Weather forecast channel

# Get OnePlus weather
app_category("net.oneplus.weather")
}
