% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ccopy}
\alias{ccopy}
\title{Copy mpathsenser zip files to a new location}
\usage{
ccopy(from, to, recursive = TRUE)
}
\arguments{
\item{from}{A path to copy files from.}

\item{to}{A path to copy files to.}

\item{recursive}{Should files from subdirectories be copied?}
}
\value{
A message indicating how many files were copied.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Copy zip files from a source destination to an origin destination where they do not yet exist.
That is, it only updates the target folder from the source folder.
}
\examples{
\dontrun{
ccopy("K:/data/myproject/", "~/myproject")
}
}
