% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{close_db}
\alias{close_db}
\title{Close a database connection}
\usage{
close_db(db)
}
\arguments{
\item{db}{A database connection to an m-Path Sense database.}
}
\value{
Returns invisibly regardless of whether the database is active, valid,
or even exists.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

This is a convenience function that is simply a wrapper around \code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}}.
}
\seealso{
\code{\link[=open_db]{open_db()}} for opening an mpathsenser database.
}
