% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{close_db}
\alias{close_db}
\title{Close a database connection}
\usage{
close_db(db)
}
\arguments{
\item{db}{A database connection}
}
\description{
This is a convenience function that is simply a wrapper around \link[DBI]{dbDisconnect}.
}
\details{
This function returns invisibly regardless of whether the database is active, valid, or even
exists.
}
