% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{copy_db}
\alias{copy_db}
\title{Copy (a subset of) a database to another database}
\usage{
copy_db(source_db, target_db, sensor = "All")
}
\arguments{
\item{source_db}{A mpathsenser database connection from where the data will be transferred.}

\item{target_db}{A mpathsenser database connection where the data will be transferred to.
\code{\link[=create_db]{create_db()}} to create a new database.}

\item{sensor}{A character vector containing one or multiple sensors. See
\code{\link[mpathsenser]{sensors}} for a list of available sensors. Use "All" for all available
sensors.}
}
\value{
No return value, called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
