% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coverage}
\alias{coverage}
\title{Create a coverage chart showing sampling rate}
\usage{
coverage(data, frequency)
}
\arguments{
\item{data}{A list of data frames containing data per sensor
as output by \link[CARP]{import}}

\item{frequency}{A named numeric vector with sensors as names
and the number of expected samples per hour}
}
\value{

}
\description{
Only applicable to non-reactive sensors with "continuous" sampling
}
\examples{
setwd("~/data")
clean()
data <- import()
freq <- c(
  accelerometer = 3600, # Once per second
  air_quality = 1,
  app_usage = 360, # Once per 10 seconds
  bluetooth = 60,  # Once per minute
  light = 360, # Once per 10 seconds
  location = 120, # Once per 30 seconds
  memory = 60, # Once per minute
  noise = 120,
  weather = 1,
  wifi = 60 # once per minute
)
coverage(data, freq)
}
