% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coverage}
\alias{coverage}
\title{Create a coverage chart showing sampling rate}
\usage{
coverage(
  db,
  participant_id,
  sensor = "All",
  frequency,
  relative = TRUE,
  offset = "None",
  startDate = NULL,
  endDate = NULL
)
}
\arguments{
\item{frequency}{A named numeric vector with sensors as names
and the number of expected samples per hour}

\item{data}{A valid database connection. Schema must be that as it is created by
\link[CARP]{open_db}.}
}
\value{

}
\description{
Only applicable to non-reactive sensors with "continuous" sampling
}
\examples{
setwd("~/data")
clean()
import()
freq <- c(
  accelerometer = 3600, # Once per second
  air_quality = 1,
  app_usage = 360, # Once per 10 seconds
  bluetooth = 60,  # Once per minute
  light = 360, # Once per 10 seconds
  location = 120, # Once per 30 seconds
  memory = 60, # Once per minute
  noise = 120,
  weather = 1,
  wifi = 60 # once per minute
)
coverage(db, freq)
}
