% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_db}
\alias{create_db}
\title{Create a new mpathsenser database}
\usage{
create_db(path = getwd(), db_name = "sense.db", overwrite = FALSE)
}
\arguments{
\item{path}{The path to the database.}

\item{db_name}{The name of the database.}

\item{overwrite}{In case a database with \code{db_name} already exists, indicate whether it should
be overwritten or not. Otherwise, this option is ignored.}
}
\value{
A database connection using prepared database schemas.
}
\description{
Create a new mpathsenser database
}
