% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{create_db}
\alias{create_db}
\title{Create a new mpathsenser database}
\usage{
create_db(path = getwd(), db_name = "sense.db", overwrite = FALSE)
}
\arguments{
\item{path}{The path to the database.}

\item{db_name}{The name of the database.}

\item{overwrite}{In case a database with \code{db_name} already exists, indicate whether it should
be overwritten or not. Otherwise, this option is ignored.}
}
\value{
A database connection using prepared database schemas.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
# Create a new database in a temporary directory
db <- create_db(tempdir(), "mydb.db")

# You can also create an in-memory database
db2 <- create_db(path = NULL, ":memory:")

# Cleanup
close_db(db)
close_db(db2)
file.remove(file.path(tempdir(), "mydb.db"))
}
