% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{decrypt_gps}
\alias{decrypt_gps}
\title{Decrypt GPS data from a curve25519 public key}
\usage{
decrypt_gps(data, key)
}
\arguments{
\item{data}{A (lazy) tibble containing the GPS data}

\item{key}{A curve25519 public key}
}
\value{
The non-lazy decrypted GPS data
}
\description{
Decrypt GPS data from a curve25519 public key
}
