% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_jsons}
\alias{fix_jsons}
\title{Fix end-of-line of JSON files}
\usage{
fix_jsons(path = getwd(), files = NULL, parallel = FALSE)
}
\arguments{
\item{path}{The path name of the JSON files.}

\item{files}{Alternatively, a character list of the input files}

\item{parallel}{A logical value whether you want to check in parallel. Useful for a lot of files.}
}
\value{
A message indicating whether any files need to be fixed.
}
\description{
When copying data directly coming from CARP, JSON files are sometimes corrupted due to the
app not properly closing them. This function attempts to fix the most common
problems associated with improper file closure by CARP.
}
\examples{
\dontrun{
future::plan(future::multisession)
files <- test_jsons()
fix_jsons(files = files)
}
}
