% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{freq}
\alias{freq}
\title{Measurement frequencies per sensor}
\usage{
freq
}
\value{
This vector contains the following
information:\tabular{lll}{
   Sensor \tab Frequency (per hour) \tab Full text \cr
   Accelerometer \tab 720 \tab Once per 5 seconds. Can have multiple instances. \cr
   AirQuality \tab 1 \tab Once per hour. \cr
   AppUsage \tab 2 \tab Once every 30 minutes. Can have multiple instances. \cr
   Bluetooth \tab 12 \tab Once every 5 minutes. Can have multiple instances. \cr
   Gyroscope \tab 720 \tab Once per 5 seconds. Can have multiple instances. \cr
   Light \tab 360 \tab Once per 10 seconds. \cr
   Location \tab 60 \tab Once every 60 seconds. \cr
   Memory \tab 60 \tab Once per minute \cr
   Noise \tab 120 \tab Once every 30 seconds. Microphone cannot be used in the background in Android 11. \cr
   Weather \tab 1 \tab Once per hour. \cr
   Wifi \tab 60 \tab Once per minute. \cr
}
}
\description{
A numeric vector containing (an example) of example measurement frequencies per sensor.
Such input is needed for \code{\link[=coverage]{coverage()}}.
}
\examples{
freq
}
