% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{data}
\name{freq}
\alias{freq}
\title{Measurement frequencies per sensor}
\format{
An object of class \code{numeric} of length 12.
}
\usage{
freq
}
\description{
A numeric vector containing (an example) of the measurement frequencies per sensor.
Such input is needed for \link[CARP]{coverage}.
}
\keyword{datasets}
