% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_nrows}
\alias{get_nrows}
\title{Get the number of rows per sensor in a mpathsenser database}
\usage{
get_nrows(
  db,
  sensor = NULL,
  participant_id = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{db}{db A database connection, as created by \code{\link[=create_db]{create_db()}}.}

\item{sensor}{A character vector of one or multiple vectors. Use \code{sensor = NULL} for all
sensors. See \link[mpathsenser]{sensors} for a list of all available sensors.}

\item{participant_id}{A single participant identifier (stored as an unsigned integer; an
integer, numeric, or character value is accepted). Use
\code{\link[=get_participants]{get_participants()}} to retrieve all participants from the database. Leave empty to get data
for all participants.}

\item{start_date}{Optional search window specifying date where to begin search. Must be
convertible to date using \code{\link[base:as.Date]{base::as.Date()}}. Use \code{\link[=first_date]{first_date()}} to find the date of the first
entry for a participant.}

\item{end_date}{Optional search window specifying date where to end search. Must be convertible
to date using \code{\link[base:as.Date]{base::as.Date()}}. Use \code{\link[=last_date]{last_date()}} to find the date of the last entry for a
participant.}
}
\value{
A named vector containing the number of rows for each sensor.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
\dontrun{
# Open a database connection
db <- open_db("path/to/db")

# Get the number of rows for all sensors
get_nrows(db, sensor = NULL)

# Get the number of rows for the Accelerometer and Gyroscope sensors
get_nrows(db, c("Accelerometer", "Gyroscope"))

# Remember to close the connection
close_db(db)
}
}
