% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_nrows}
\alias{get_nrows}
\title{Get the number of rows sensors in a CARP database}
\usage{
get_nrows(
  db,
  sensor = "All",
  participant_id = NULL,
  start_date = NULL,
  end_date = NULL
)
}
\arguments{
\item{db}{db A database connection, as created by \link[CARP]{create_db}.}

\item{sensor}{A character vector of one or multiple vectors. Use "All" for all sensors. See
\link[CARP]{sensors} for a list of all available sensors.}
}
\value{
A named vector containing the number of rows for each sensor.
}
\description{
Get the number of rows sensors in a CARP database
}
