% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_participants}
\alias{get_participants}
\title{Get all participants}
\usage{
get_participants(db, lazy = FALSE)
}
\arguments{
\item{db}{db A database connection, as created by \link[mpathsenser]{create_db}.}

\item{lazy}{Whether to evaluate lazily using \link[dbplyr]{dbplyr}.}
}
\value{
A data frame containing all \code{participant_id} and \code{study_id}.
}
\description{
Get all participants
}
