% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_processed_files}
\alias{get_processed_files}
\title{Get all processed files from a database}
\usage{
get_processed_files(db)
}
\arguments{
\item{db}{A database connection, as created by \code{\link[=create_db]{create_db()}}.}
}
\value{
A data frame containing the \code{file_name}, \code{participant_id}, and \code{study_id} of the
processed files.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
