% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_processed_files}
\alias{get_processed_files}
\title{Get all processed files from a database}
\usage{
get_processed_files(db)
}
\arguments{
\item{db}{A database connection, as created by \link[CARP]{create_db()}.}
}
\value{
A data frame contain processed file for each participant and study.
}
\description{
Get all processed files from a database
}
