% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_studies}
\alias{get_studies}
\title{Get all studies}
\usage{
get_studies(db, lazy = FALSE)
}
\arguments{
\item{db}{db A database connection, as created by \link[mpathsenser]{create_db}.}

\item{lazy}{Whether to evaluate lazily using \link[dbplyr]{dbplyr}.}
}
\value{
A data frame containing all studies.
}
\description{
Get all studies
}
