% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_studies}
\alias{get_studies}
\title{Get all studies}
\usage{
get_studies(db, lazy = FALSE)
}
\arguments{
\item{db}{db A database connection, as created by \code{\link[=create_db]{create_db()}}.}

\item{lazy}{Whether to evaluate lazily using \link[dbplyr]{dbplyr}.}
}
\value{
A data frame containing all studies.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
# Create a database
db <- create_db(tempdir(), "mydb.db")

# Add some studies
DBI::dbExecute(db, "INSERT INTO Study VALUES('study1', 'data_format1')")

# Get the studies
get_studies(db)

# Cleanup
close_db(db)
file.remove(file.path(tempdir(), "mydb.db"))
}
