% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_functions.R
\name{haversine}
\alias{haversine}
\title{Calculate the Great-Circle Distance between two points in meters}
\usage{
haversine(lon1, lat1, lon2, lat2, r = 6371)
}
\arguments{
\item{lon1}{The longitude of point 1 in degrees.}

\item{lat1}{The latitude of point 1 in degrees.}

\item{lon2}{The longitude of point 2 in degrees.}

\item{lat2}{The latitude of point 2 in degrees.}

\item{r}{The average earth radius (6371 km by default).}
}
\value{
The distance between point 1 and 2 in kilometers.
}
\description{
Calculate the great-circle distance between two points using the Haversine function.
}
