% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import}
\alias{import}
\title{Import CARP files into a database (CARP data scheme)}
\usage{
import(
  path = getwd(),
  db = NULL,
  dbname = "carp.db",
  overwrite_db = TRUE,
  backend = "RSQLite",
  progress = TRUE,
  recursive = TRUE,
  parallel = FALSE
)
}
\arguments{
\item{path}{The path to the file directory}

\item{db}{Valid database connection.}

\item{dbname}{If no database is provided, a new database dbname is created.}

\item{overwrite_db}{If a database with the same \code{dbname}  already exists, should it be
overwritten?}

\item{backend}{Name of the database backend that is used. Currently, only RSQLite is supported.}

\item{progress}{Logical value to show a progress bar or not.}

\item{recursive}{Should the listing recurse into directories?}

\item{parallel}{A value that indicates whether to do reading in and processing
in parallel. If this argument is a number, this indicates the number of workers that will be
used.}
}
\value{
Invisible. Imported database can be reopened using \link[CARP]{open_db}.
}
\description{
Currently, only SQLite is supported as a backend. Due to its concurrency restriction, the
\code{parallel} option is disabled.
}
