% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import}
\alias{import}
\title{Import CARP files into R (CARP data scheme)}
\usage{
import(path = getwd(), progress = TRUE, parallel = TRUE)
}
\arguments{
\item{path}{The path to the file directory}

\item{progress}{Logical value to show a progress bar or not.}

\item{parallel}{Logical value which indicates whether to do reading in and processing in parallel.}
}
\value{
An S3 object containing a tibble for each sensor in the data.
}
\description{
Import CARP files into R (CARP data scheme)
}
