% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{index_db}
\alias{index_db}
\title{Create indexes for an mpathsenser database}
\usage{
index_db(db)
}
\arguments{
\item{db}{A database connection to an m-Path Sense database.}
}
\value{
No return value, called for side effects.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
