% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensor_functions.R
\name{moving_average}
\alias{moving_average}
\title{Moving average for values in CARP DB}
\usage{
moving_average(
  db,
  sensor,
  participant_id,
  ...,
  n,
  startDate = NULL,
  endDate = NULL
)
}
\arguments{
\item{db}{A database connection to a CARP database.}

\item{sensor}{The name of a sensor. See \link[CARP]{sensors} for a list of available sensors.}

\item{participant_id}{A character string identifying a single participant. Use
\code{\link[CARP]{get_participants}} to retrieve all participants from the database.
Leave empty to get data for all participants.}

\item{...}{Unquoted names of columns of the \code{sensor} table.}

\item{n}{The number of observations to average over.}

\item{startDate}{Optional search window specifying date where to begin search. Must be convertible to date using \link[base]{as.Date}. Use \link[CARP]{first_date} to find the date of the first entry for a participant.}

\item{endDate}{Optional search window specifying date where to end search. Must be convertible to date using \link[base]{as.Date}. Use \link[CARP]{last_date} to find the date of the last entry for a participant.}
}
\value{
A tibble with the same columns as the input, modified to be a moving average.
}
\description{
Moving average for values in CARP DB
}
\examples{
\dontrun{
get_moving_average(db, "Light", "27624", mean_lux, max_lux, n = 5)
}
}
