% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpathsenser-package.R
\docType{package}
\name{mpathsenser-package}
\alias{mpathsenser}
\alias{mpathsenser-package}
\title{mpathsenser: Process and Analyse Data from m-Path Sense}
\description{
Overcomes one of the major challenges in mobile (passive) sensing, namely
being able to pre-process the raw data that comes from a mobile sensing app,
specifically "m-Path Sense" \url{https://m-path.io}. The main task of 'mpathsenser' is
therefore to read "m-Path Sense" JSON files into a database and provide several
convenience functions to aid in data processing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitlab.kuleuven.be/ppw-okpiv/researchers/u0134047/mpathsenser/}
  \item \url{https://ppw-okpiv.pages.gitlab.kuleuven.be/researchers/u0134047/mpathsenser/}
  \item Report bugs at \url{https://gitlab.kuleuven.be/ppw-okpiv/researchers/u0134047/mpathsenser/-/issues/}
}

}
\author{
\strong{Maintainer}: Koen Niemeijer \email{koen.niemeijer@kuleuven.be} (\href{https://orcid.org/0000-0002-0816-534X}{ORCID})

Other contributors:
\itemize{
  \item Kristof Meers \email{kristof.meers@kuleuven.be} (\href{https://orcid.org/0000-0002-9610-7712}{ORCID}) [contributor]
  \item KU Leuven [copyright holder, funder]
}

}
\keyword{internal}
