% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpathsenser-package.R
\docType{package}
\name{mpathsenser}
\alias{mpathsenser}
\title{mpathsenser: A package for processing and analysing data from m-Path Sense}
\description{
foo, bar, description
}
\section{Foo functions}{

foo, bar
}

\section{Author}{

\strong{Maintainer}: Koen Niemeijer \email{koen.niemeijer@kuleuven.be}
(\href{https://orcid.org/0000-0002-0816-534X}{ORCID})

Other contributors:
\itemize{
\item KU Leuven [copyright holder, funder]
}
}

