% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{open_db}
\alias{open_db}
\title{Open an mpathsenser database.}
\usage{
open_db(path = getwd(), db_name = "sense.db")
}
\arguments{
\item{path}{The path to the database. Use NULL to use the full path name in db_name.}

\item{db_name}{The name of the database.}
}
\value{
A connection to an mpathsenser database.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\seealso{
\code{\link[mpathsenser]{close_db}} for closing a database;
\code{\link[mpathsenser]{copy_db}} for copying (part of) a database;
\code{\link[mpathsenser]{index_db}} for indexing a database;
\code{\link[mpathsenser]{get_data}} for extracting data from a database.
}
