% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{open_db}
\alias{open_db}
\title{Open a CARP database}
\usage{
open_db(db_name = "carp.db")
}
\arguments{
\item{db_name}{The name of the database.}
}
\value{
A database connection.
}
\description{
Open a CARP database
}
