% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{open_db}
\alias{open_db}
\title{Open a CARP database}
\usage{
open_db(path = getwd(), db_name = "carp.db")
}
\arguments{
\item{path}{The path to the database. Use NULL to use the full path name in db_name.}

\item{db_name}{The name of the database.}
}
\value{
A database connection.
}
\description{
Open a CARP database
}
