% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage.R
\name{plot.coverage}
\alias{plot.coverage}
\title{Plot a coverage overview}
\usage{
\method{plot}{coverage}(x, ...)
}
\arguments{
\item{x}{A tibble with the coverage data coming from \code{\link[mpathsenser]{coverage}}.}

\item{...}{Other arguments passed on to methods. Not currently used.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Plot a coverage overview
}
\seealso{
\code{\link[mpathsenser]{coverage}}
}
\keyword{internal}
