% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_jsons}
\alias{test_jsons}
\title{Test JSON files for being in the correct format.}
\usage{
test_jsons(path = getwd(), parallel = FALSE)
}
\arguments{
\item{path}{The pathname of the JSON files.}

\item{parallel}{A logical value whether you want to check in parallel. Useful when there are a lot of files. If you have already used \code{future::plan}, you can leave this parameter to \code{FALSE}.}
}
\value{
A message indicating whether there were any issues and a character vector of the file names that need to be fixed. If there were no issues, no result is returned.
}
\description{
Test JSON files for being in the correct format.
}
