% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{test_jsons}
\alias{test_jsons}
\title{Test JSON files for being in the correct format.}
\usage{
test_jsons(path = getwd(), parallel = FALSE)
}
\arguments{
\item{path}{The pathname of the JSON files.}

\item{parallel}{A logical value whether you want to check in parallel. Useful for a lot of files.}
}
\value{
A message indicating if there are any files to fixes.
}
\description{
Test JSON files for being in the correct format.
}
