% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{vacuum_db}
\alias{vacuum_db}
\title{Vacuum a database}
\usage{
vacuum_db(db, ...)
}
\arguments{
\item{db}{A database connection to an m-Path Sense database.}

\item{...}{Currently not used.}
}
\value{
An integer indicating the number of rows that were vacuumed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

The SQL \code{VACUUM} command reclaims storage occupied by dead rows. This is mostly useful for
SQLite databases, as the \code{VACUUM} command rebuilds the database file, repacking it into a minimal
amount of disk space. Note that MySQL databases need not (nor can) be vacuumed.
}
